pyproject.toml
[build-system]
requires = ["setuptools >= 42", "versioningit ~= 1.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atomate2-turbomole"
description = "The atomate2-turbomole package is a workflow package for Turbomole"
readme = "README.md"
keywords = ["high-throughput", "automated", "workflow", "dft", "ab initio", "turbomole"]
license = { text = "modified BSD" }
authors = [{ name = "David Waroquiers", email = "david.waroquiers@matgenix.com" }]
dynamic = ["version"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology",
    "Operating System :: OS Independent",
    "Topic :: Other/Nonlisted Topic",
    "Topic :: Scientific/Engineering",
]
requires-python = ">=3.8"
dependencies = [
    "pymatgen>=2022.2.10",
    "custodian>=2024.3.12",
    "jobflow>=0.1.11",
    "emmet-core>=0.80.0",
    "turbomoleio==1.4.0"
]

[project.optional-dependencies]
dev = ["pre-commit>=2.12.1"]
tests = ["pytest==7.2.2", "pytest-cov==4.0.0", "pytest-mock==3.6.1", "mongomock==4.1.2"]

[tool.versioningit.vcs]
method = "git"
default-tag = "0.0.1"

[tool.flake8]
max-line-length = 88
max-doc-length = 88
select = "C, E, F, W, B"
extend-ignore = "E203, W503, E501, F401, RST21"
min-python-version = "3.8.0"
docstring-convention = "numpy"
rst-roles = "class, func, ref, obj"

[tool.mypy]
ignore_missing_imports = true
no_strict_optional = true

[tool.pytest.ini_options]
addopts = "--strict-markers --cov=atomate2.turbomole --cov-config=pyproject.toml --cov-report=html --cov-report=term"
filterwarnings = []

[tool.coverage.run]
include = ["src/*"]
parallel = true
branch = true

[tool.coverage.paths]
source = ["src/"]

[tool.coverage.report]
skip_covered = true
show_missing = true
exclude_lines = [
    '^\s*assert False(,|$)',
    'if typing.TYPE_CHECKING:',
    '^\s*@overload( |$)',
    'pragma: no cover',
    'NotImplementedError',
]
partial_branches = [
    'pragma: no branch',
]
fail_under = 100

[tool.ruff]
target-version = "py38"
select = [
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "D",   # pydocstyle
    "E",   # pycodestyle
    "F",   # pyflakes
    "I",   # isort
    "PLE", # pylint error
    "PLW", # pylint warning
    "Q",   # flake8-quotes
    "RUF", # Ruff-specific rules
    "SIM", # flake8-simplify
    "TID", # tidy imports
    "UP",  # pyupgrade
    "W",   # pycodestyle
    "YTT", # flake8-2020
]
pydocstyle.convention = "numpy"
isort.known-first-party = ["atomate2.turbomole"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"**/tests/*" = ["D"]

[tool.setuptools.package-data]
"atomate2.turbomole.schemas.calc_types" = ["*.yaml"]

#
#
#
#
#[tool.poetry]
#name = "turbomole"
#version = "0.3.1"
#description = "The turbomole package is a workflow package for Toyota Motor Europe"
#authors = ["davidwaroquiers <david.waroquiers@matgenix.com>"]
#packages = [
#  { include = "turbomole", from = "src" }
#]
#
#[tool.poetry.dependencies]
#python = "^3.8"
#custodian = "^2021.2.8"
#jobflow = "^0.1.2"
#fireworks = "^2.0.3"
#
#[tool.poetry.dev-dependencies]
#pytest = "^5.2"
#pytest-cov = "^2.10.1"
#coverage = {version = "^5.3", extras = ["toml"]}
#pytest-mock = "^3.3.1"
#pre-commit = "^2.9.3"
#isort = "^5.6.4"
#black = "^20.8b1"
#flake8 = "^3.8.4"
#pylint = "^2.6.0"
#pydocstyle = "^6.1.1"
#mypy = "^0.790"
#
#[build-system]
#requires = ["poetry-core>=1.0.0"]
#build-backend = "poetry.core.masonry.api"
#
#[tool.isort]
#profile = "black"
#
#[tool.pydocstyle]
#inherit = false
#
#[tool.pytest.ini_options]
#minversion = "6.1.2"
#testpaths = "tests"
##addopts = "--strict-markers --cov=src/turbomole -cov-config=pyproject.toml --cov-report html --cov-report term"
#
## Coverage options
#[tool.coverage.run]
#source = ["turbomole"]
#branch = true
#
#[tool.coverage.report]
#exclude_lines = [
#    "pragma: no cover",
#    "raise NotImplementedError",
#]
#fail_under = 100
#
#[tool.coverage.html]
#directory = "htmlcov"
